requirements.txt
dicompyler-core>=0.5.4
wxPython>=4.0.4
pypubsub
numpy
scipy
pydicom>1.0
bokeh>=1.2.0
psycopg2-binary
shapely[vectorized]<1.7.0
python-dateutil
statsmodels
scikit-learn==0.21
regressors